[SPARK-18287][SQL] Move hash expressions from misc.scala into hash.scala
